Migrate fjs/effects/eff to .f.mjs, split Eff type into types.ts
proof.f.ts stays TypeScript for now: it has a runtime dependency on
../proof.f.ts (fjs/effects/proof.f.ts), still unmigrated. Updates 4
dependents (protocol/mcp, cas, cas/evo, emergent_testing) and 3 doc
references.
