+++ fjs/emergent_testing/module.f.ts
@@ -31,7 +31,7 @@ import {
 } from '../effects/node/module.f.ts'
 import { history, historyStep, pure, step } from '../effects/module.f.mjs'
 import type { Effect, Operation } from '../effects/types.ts'
-import { eff } from '../effects/eff/module.f.ts'
+import { eff } from '../effects/eff/module.f.mjs'
 import { loadModuleMap, shouldLoad, type LoadModuleOperations, type ModuleMap } from '../dev/module.f.ts'
 import { invert } from '../types/result/module.f.mjs'
 import { definedEntries } from '../types/object/module.f.mjs'
